Rebase to Turtlecoin codebase
- Bug fixes and improvements
- v0.0.3

+++ src/WalletBackend/WalletBackend.cpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2018-2019, The TurtleCoin Developers
-// 
+//
 // Please see the included LICENSE file for more information.
 
 ////////////////////////////////////////
@@ -11,11 +11,8 @@
 
 #include <config/CryptoNoteConfig.h>
 
+#include <crypto/crypto.h>
 #include <crypto/random.h>
-
-#include <CryptoNoteCore/Account.h>
-#include <CryptoNoteCore/CryptoNoteTools.h>
-#include <CryptoNoteCore/CryptoNoteBasicImpl.h>
 
 #include <cryptopp/aes.h>
 #include <cryptopp/algparam.h>
@@ -33,6 +30,8 @@
 #include <iomanip>
 
 #include "JsonHelper.h"
+
+#include <Logger/Logger.h>
 
 #include <Mnemonics/Mnemonics.h>
 
@@ -97,7 +96,7 @@ Error checkNewWalletFilename(std::string filename)
 
     /* Don't leave random files around if we fail later down the road */
     fs::remove(filename);
-    
+
     return SUCCESS;
 }
 
@@ -112,7 +111,7 @@ WalletBackend::WalletBackend()
 {
     m_eventHandler = std::make_shared<EventHandler>();
 
-    /* Remember to correctly initialize the daemon - 
+    /* Remember to correctly initialize the daemon -
     we can't do it here since we don't have the host/port, and the json
     serialization uses the default constructor */
 }
@@ -137,11 +136,12 @@ WalletBackend::WalletBackend(
     const uint64_t scanHeight,
     const bool newWallet,
     const std::string daemonHost,
-    const uint16_t daemonPort) :
+    const uint16_t daemonPort,
+    const bool daemonSSL) :
 
     m_filename(filename),
     m_password(password),
-    m_daemon(std::make_shared<Nigel>(daemonHost, daemonPort))
+    m_daemon(std::make_shared<Nigel>(daemonHost, daemonPort, daemonSSL))
 {
     /* Generate the address from the two private keys */
     std::string address = Utilities::privateKeysToAddress(
@@ -163,11 +163,12 @@ WalletBackend::WalletBackend(
     const std::string address,
     const uint64_t scanHeight,
     const std::string daemonHost,
-    const uint16_t daemonPort) :
+    const uint16_t daemonPort,
+    const bool daemonSSL) :
 
     m_filename(filename),
     m_password(password),
-    m_daemon(std::make_shared<Nigel>(daemonHost, daemonPort))
+    m_daemon(std::make_shared<Nigel>(daemonHost, daemonPort, daemonSSL))
 {
     bool newWallet = false;
 
@@ -182,43 +183,6 @@ WalletBackend::WalletBackend(
 /* STATIC FUNCTIONS */
 //////////////////////
 
-std::tuple<Error, std::string> WalletBackend::createIntegratedAddress(
-    const std::string address,
-    const std::string paymentID)
-{
-    if (Error error = validatePaymentID(paymentID); error != SUCCESS)
-    {
-        return {error, std::string()};
-    }
-
-    const bool allowIntegratedAddresses = false;
-
-    if (Error error = validateAddresses({address}, allowIntegratedAddresses); error != SUCCESS)
-    {
-        return {error, std::string()};
-    }
-
-    uint64_t prefix;
-
-    CryptoNote::AccountPublicAddress addr;
-
-    /* Get the private + public key from the address */
-    CryptoNote::parseAccountAddressString(prefix, addr, address);
-
-    /* Pack as a binary array */
-    CryptoNote::BinaryArray ba;
-    CryptoNote::toBinaryArray(addr, ba);
-    std::string keys = Common::asString(ba);
-
-    /* Encode prefix + paymentID + keys as an address */
-    const std::string integratedAddress = Tools::Base58::encode_addr(
-        CryptoNote::parameters::CRYPTONOTE_PUBLIC_ADDRESS_BASE58_PREFIX,
-        paymentID + keys
-    );
-
-    return {SUCCESS, integratedAddress};
-}
-
 /* Imports a wallet from a mnemonic seed. Returns the wallet class,
    or an error. */
 std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importWalletFromSeed(
@@ -227,12 +191,13 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importWalletFro
     const std::string password,
     const uint64_t scanHeight,
     const std::string daemonHost,
-    const uint16_t daemonPort)
+    const uint16_t daemonPort,
+    const bool daemonSSL)
 {
     /* Check the filename is valid */
     if (Error error = checkNewWalletFilename(filename); error != SUCCESS)
     {
-        return {error, nullptr}; 
+        return {error, nullptr};
     }
 
     /* Convert the mnemonic into a private spend key */
@@ -240,15 +205,20 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importWalletFro
 
     if (mnemonicError)
     {
-        return {mnemonicError, nullptr}; 
+        return {mnemonicError, nullptr};
     }
 
     Crypto::SecretKey privateViewKey;
 
     /* Derive the private view key from the private spend key */
-    CryptoNote::AccountBase::generateViewFromSpend(
+    Crypto::crypto_ops::generateViewFromSpend(
         privateSpendKey, privateViewKey
     );
+
+    if (Error error = validatePrivateKey(privateViewKey); error != SUCCESS)
+    {
+        return {error, nullptr};
+    }
 
     /* Just defining here so it's more obvious what we're doing in the
        constructor */
@@ -256,7 +226,7 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importWalletFro
 
     const std::shared_ptr<WalletBackend> wallet(new WalletBackend(
         filename, password, privateSpendKey, privateViewKey,
-        scanHeight, newWallet, daemonHost, daemonPort
+        scanHeight, newWallet, daemonHost, daemonPort, daemonSSL
     ));
 
     wallet->init();
@@ -276,12 +246,23 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importWalletFro
     const std::string password,
     const uint64_t scanHeight,
     const std::string daemonHost,
-    const uint16_t daemonPort)
+    const uint16_t daemonPort,
+    const bool daemonSSL)
 {
     /* Check the filename is valid */
     if (Error error = checkNewWalletFilename(filename); error != SUCCESS)
     {
-        return {error, nullptr}; 
+        return {error, nullptr};
+    }
+
+    if (Error error = validatePrivateKey(privateViewKey); error != SUCCESS)
+    {
+        return {error, nullptr};
+    }
+
+    if (Error error = validatePrivateKey(privateSpendKey); error != SUCCESS)
+    {
+        return {error, nullptr};
     }
 
     /* Just defining here so it's more obvious what we're doing in the
@@ -290,7 +271,7 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importWalletFro
 
     const std::shared_ptr<WalletBackend> wallet(new WalletBackend(
         filename, password, privateSpendKey, privateViewKey, scanHeight,
-        newWallet, daemonHost, daemonPort
+        newWallet, daemonHost, daemonPort, daemonSSL
     ));
 
     wallet->init();
@@ -310,17 +291,30 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importViewWalle
     const std::string password,
     const uint64_t scanHeight,
     const std::string daemonHost,
-    const uint16_t daemonPort)
+    const uint16_t daemonPort,
+    const bool daemonSSL)
 {
     /* Check the filename is valid */
     if (Error error = checkNewWalletFilename(filename); error != SUCCESS)
     {
-        return {error, nullptr}; 
+        return {error, nullptr};
+    }
+
+    if (Error error = validatePrivateKey(privateViewKey); error != SUCCESS)
+    {
+        return {error, nullptr};
+    }
+
+    const bool allowIntegratedAddresses = false;
+
+    if (Error error = validateAddresses({address}, allowIntegratedAddresses); error != SUCCESS)
+    {
+        return {error, nullptr};
     }
 
     const std::shared_ptr<WalletBackend> wallet(new WalletBackend(
         filename, password, privateViewKey, address, scanHeight, daemonHost,
-        daemonPort
+        daemonPort, daemonSSL
     ));
 
     wallet->init();
@@ -336,14 +330,15 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::createWallet(
     const std::string filename,
     const std::string password,
     const std::string daemonHost,
-    const uint16_t daemonPort)
+    const uint16_t daemonPort,
+    const bool daemonSSL)
 {
     /* Check the filename is valid */
     if (Error error = checkNewWalletFilename(filename); error != SUCCESS)
     {
         return {error, nullptr};
     }
-	
+
     CryptoNote::KeyPair spendKey;
     Crypto::SecretKey privateViewKey;
     Crypto::PublicKey publicViewKey;
@@ -352,7 +347,7 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::createWallet(
     Crypto::generate_keys(spendKey.publicKey, spendKey.secretKey);
 
     /* Derive the view key from the spend key */
-    CryptoNote::AccountBase::generateViewFromSpend(
+    Crypto::crypto_ops::generateViewFromSpend(
         spendKey.secretKey, privateViewKey, publicViewKey
     );
 
@@ -363,9 +358,9 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::createWallet(
 
     const std::shared_ptr<WalletBackend> wallet(new WalletBackend(
         filename, password, spendKey.secretKey, privateViewKey,
-        scanHeight, newWallet, daemonHost, daemonPort
+        scanHeight, newWallet, daemonHost, daemonPort, daemonSSL
     ));
-	
+
     wallet->init();
 
     /* Save to disk */
@@ -379,7 +374,8 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::openWallet(
     const std::string filename,
     const std::string password,
     const std::string daemonHost,
-    const uint16_t daemonPort)
+    const uint16_t daemonPort,
+    const bool daemonSSL)
 {
     /* Open in binary mode, since we have encrypted data */
     std::ifstream file(filename, std::ios_base::binary);
@@ -387,7 +383,7 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::openWallet(
     /* Check we successfully opened the file */
     if (!file)
     {
-        return {FILENAME_NON_EXISTENT, nullptr}; 
+        return {FILENAME_NON_EXISTENT, nullptr};
     }
 
     /* Read file into a buffer */
@@ -403,7 +399,7 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::openWallet(
 
     if (error)
     {
-        return {error, nullptr}; 
+        return {error, nullptr};
     }
 
     using namespace CryptoPP;
@@ -467,7 +463,7 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::openWallet(
 
     if (error)
     {
-        return {error, nullptr}; 
+        return {error, nullptr};
     }
 
     try
@@ -494,13 +490,19 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::openWallet(
         /* Initialize it from the json (We could do this in less steps, but it
            requires a move/copy constructor) */
         error = wallet->fromJSON(
-            walletJson, filename, password, daemonHost, daemonPort
+            walletJson, filename, password, daemonHost, daemonPort, daemonSSL
         );
 
         return {error, wallet};
     }
     catch (const std::invalid_argument &e)
     {
+        Logger::logger.log(
+            std::string("Failed to open wallet file: ") + e.what(),
+            Logger::FATAL,
+            {Logger::FILESYSTEM, Logger::SAVE}
+        );
+
         return {WALLET_FILE_CORRUPTED, nullptr};
     }
 }
@@ -525,7 +527,7 @@ void WalletBackend::init()
         auto [startHeight, startTimestamp] = m_subWallets->getMinInitialSyncStart();
 
         m_walletSynchronizer = std::make_shared<WalletSynchronizer>(
-            m_daemon, 
+            m_daemon,
             startHeight,
             startTimestamp,
             m_subWallets->getPrivateViewKey(),
@@ -538,7 +540,7 @@ void WalletBackend::init()
         m_walletSynchronizer->initializeAfterLoad(m_daemon, m_eventHandler);
     }
 
-    m_walletSynchronizer->m_subWallets = m_subWallets;
+    m_walletSynchronizer->setSubWallets(m_subWallets);
 
     /* Launch the wallet sync process in a background thread */
     m_walletSynchronizer->start();
@@ -606,6 +608,12 @@ Error WalletBackend::unsafeSave() const
 
     if (!file)
     {
+        Logger::logger.log(
+            std::string("Wallet filename: ") + m_filename + " is invalid",
+            Logger::FATAL,
+            {Logger::FILESYSTEM, Logger::SAVE}
+        );
+
         return INVALID_WALLET_FILENAME;
     }
 
@@ -712,12 +720,22 @@ std::tuple<Error, Crypto::Hash> WalletBackend::sendFusionTransactionAdvanced(
 void WalletBackend::reset(uint64_t scanHeight, uint64_t timestamp)
 {
     m_syncRAIIWrapper->pauseSynchronizerToRunFunction(
-    [this, scanHeight, timestamp]() {
+    [this, scanHeight, timestamp]() mutable {
+        /* Though the wallet synchronizer can support both a timestamp and a
+           scanheight, we need a fixed scan height to cut transactions from.
+           Since a transaction in block 10 could have a timestamp before a
+           transaction in block 9, we can't rely on timestamps to reset accurately. */
+        if (timestamp != 0)
+        {
+            scanHeight = Utilities::timestampToScanHeight(timestamp);
+            timestamp = 0;
+        }
+
         /* Empty the sync status and reset the start height */
-        m_walletSynchronizer->reset(scanHeight, timestamp);
+        m_walletSynchronizer->reset(scanHeight);
 
         /* Reset transactions, inputs, etc */
-        m_subWallets->reset(scanHeight, timestamp);
+        m_subWallets->reset(scanHeight);
 
         /* Save the resetted wallet - don't need safe save, already stopped wallet
            synchronizer */
@@ -731,7 +749,7 @@ std::tuple<Error, std::string, Crypto::SecretKey> WalletBackend::addSubWallet()
 {
     return m_syncRAIIWrapper->pauseSynchronizerToRunFunction([this]() {
         /* Add the sub wallet */
-        return m_subWallets->addSubWallet(); 
+        return m_subWallets->addSubWallet();
     });
 }
 
@@ -739,11 +757,16 @@ std::tuple<Error, std::string> WalletBackend::importSubWallet(
     const Crypto::SecretKey privateSpendKey,
     const uint64_t scanHeight)
 {
+    if (Error error = validatePrivateKey(privateSpendKey); error != SUCCESS)
+    {
+        return {error, std::string()};
+    }
+
     return m_syncRAIIWrapper->pauseSynchronizerToRunFunction([&, this]() {
         /* Add the sub wallet */
         const auto [error, address] = m_subWallets->importSubWallet(
             privateSpendKey, scanHeight
-        ); 
+        );
 
         if (!error)
         {
@@ -755,10 +778,10 @@ std::tuple<Error, std::string> WalletBackend::importSubWallet(
             if (currentHeight >= scanHeight)
             {
                 /* Empty the sync status and reset the start height */
-                m_walletSynchronizer->reset(scanHeight, 0);
+                m_walletSynchronizer->reset(scanHeight);
 
                 /* Reset transactions, inputs, etc */
-                m_subWallets->reset(scanHeight, 0);
+                m_subWallets->reset(scanHeight);
             }
         }
 
@@ -770,11 +793,16 @@ std::tuple<Error, std::string> WalletBackend::importViewSubWallet(
     const Crypto::PublicKey publicSpendKey,
     const uint64_t scanHeight)
 {
+    if (Error error = validatePublicKey(publicSpendKey); error != SUCCESS)
+    {
+        return {error, std::string()};
+    }
+
     return m_syncRAIIWrapper->pauseSynchronizerToRunFunction([&, this]() {
         /* Add the sub wallet */
         const auto [error, address] = m_subWallets->importViewSubWallet(
             publicSpendKey, scanHeight
-        ); 
+        );
 
         if (!error)
         {
@@ -786,10 +814,10 @@ std::tuple<Error, std::string> WalletBackend::importViewSubWallet(
             if (currentHeight >= scanHeight)
             {
                 /* Empty the sync status and reset the start height */
-                m_walletSynchronizer->reset(scanHeight, 0);
+                m_walletSynchronizer->reset(scanHeight);
 
                 /* Reset transactions, inputs, etc */
-                m_subWallets->reset(scanHeight, 0);
+                m_subWallets->reset(scanHeight);
             }
         }
 
@@ -799,6 +827,13 @@ std::tuple<Error, std::string> WalletBackend::importViewSubWallet(
 
 Error WalletBackend::deleteSubWallet(const std::string address)
 {
+    const bool allowIntegratedAddresses = false;
+
+    if (Error error = validateAddresses({address}, allowIntegratedAddresses); error != SUCCESS)
+    {
+        return error;
+    }
+
     return m_syncRAIIWrapper->pauseSynchronizerToRunFunction([&, this]() {
         return m_subWallets->deleteSubWallet(address);
     });
@@ -865,6 +900,13 @@ Error WalletBackend::changePassword(const std::string newPassword)
 std::tuple<Error, Crypto::PublicKey, Crypto::SecretKey>
     WalletBackend::getSpendKeys(const std::string &address) const
 {
+    const bool allowIntegratedAddresses = false;
+
+    if (Error error = validateAddresses({address}, allowIntegratedAddresses); error != SUCCESS)
+    {
+        return {error, Crypto::PublicKey(), Crypto::SecretKey()};
+    }
+
     const auto [publicSpendKey, publicViewKey] = Utilities::addressToKeys(address);
 
     const auto [success, privateSpendKey] = m_subWallets->getPrivateSpendKey(publicSpendKey);
@@ -891,6 +933,13 @@ std::tuple<Error, std::string> WalletBackend::getMnemonicSeed() const
 std::tuple<Error, std::string> WalletBackend::getMnemonicSeedForAddress(
     const std::string &address) const
 {
+    const bool allowIntegratedAddresses = false;
+
+    if (Error error = validateAddresses({address}, allowIntegratedAddresses); error != SUCCESS)
+    {
+        return {error, std::string()}; 
+    }
+
     const auto privateViewKey = getPrivateViewKey();
     const auto [error, publicSpendKey, privateSpendKey] = getSpendKeys(address);
 
@@ -903,7 +952,7 @@ std::tuple<Error, std::string> WalletBackend::getMnemonicSeedForAddress(
 
     /* Derive the view key from the spend key, and check if it matches the
        actual view key */
-    CryptoNote::AccountBase::generateViewFromSpend(
+    Crypto::crypto_ops::generateViewFromSpend(
         privateSpendKey,
         derivedPrivateViewKey
     );
@@ -966,16 +1015,16 @@ std::tuple<uint64_t, std::string> WalletBackend::getNodeFee() const
     return m_daemon->nodeFee();
 }
 
-std::tuple<std::string, uint16_t> WalletBackend::getNodeAddress() const
+std::tuple<std::string, uint16_t, bool> WalletBackend::getNodeAddress() const
 {
     return m_daemon->nodeAddress();
 }
 
-void WalletBackend::swapNode(std::string daemonHost, uint16_t daemonPort)
+void WalletBackend::swapNode(std::string daemonHost, uint16_t daemonPort, bool daemonSSL)
 {
     m_syncRAIIWrapper->pauseSynchronizerToRunFunction([&, this]() {
         /* Swap and init the node */
-        m_daemon->swapNode(daemonHost, daemonPort);
+        m_daemon->swapNode(daemonHost, daemonPort, daemonSSL);
 
         /* Give the synchronizer the new daemon */
         m_walletSynchronizer->swapNode(m_daemon);
@@ -992,6 +1041,11 @@ bool WalletBackend::daemonOnline() const
 std::tuple<Error, std::string> WalletBackend::getAddress(
     const Crypto::PublicKey spendKey) const
 {
+    if (Error error = validatePublicKey(spendKey); error != SUCCESS)
+    {
+        return {error, std::string()};
+    }
+
     return m_subWallets->getAddress(spendKey);
 }
 
@@ -1057,7 +1111,8 @@ Error WalletBackend::fromJSON(
     const std::string filename,
     const std::string password,
     const std::string daemonHost,
-    const uint16_t daemonPort)
+    const uint16_t daemonPort,
+    const bool daemonSSL)
 {
     if (Error error = fromJSON(j); error != SUCCESS)
     {
@@ -1067,7 +1122,7 @@ Error WalletBackend::fromJSON(
     m_filename = filename;
     m_password = password;
 
-    m_daemon = std::make_shared<Nigel>(daemonHost, daemonPort);
+    m_daemon = std::make_shared<Nigel>(daemonHost, daemonPort, daemonSSL);
 
     init();
